chore: factor logging into logging-interface + logging packlets to break circular dep
Replace the base/loggerInterface.ts extraction approach with a proper packlet
factoring: pure-contract symbols (ILogger, IDetailLogger, ReporterLogLevel,
isDetailLogger) move to a new logging-interface packlet; logging keeps all
implementations plus shouldLog/stringifyLogValue.

collections/readOnlyConvertingResultMap.ts now imports ILogger directly from
logging-interface, breaking the cycle without type-import tricks.

The Logging.* namespace surface is exactly unchanged externally. api.md diff vs
origin/release is a true no-op (only change: removal of a spurious
ae-unresolved-link warning introduced by the prior commit).
