+++ libraries/ts-utils/src/packlets/base/index.ts
@@ -21,7 +21,6 @@
  */
 
 export * from './brand';
-export * from './loggerInterface';
 export * from './mapResults';
 export * from './messageAggregator';
 export { Normalizer } from './normalize';
