+++ libraries/ts-utils/src/packlets/collections/readOnlyConvertingResultMap.ts
@@ -20,7 +20,8 @@
  * SOFTWARE.
  */
 
-import { ILogger, captureResult, DetailedResult, failWithDetail, Result, succeedWithDetail } from '../base';
+import { captureResult, DetailedResult, failWithDetail, Result, succeedWithDetail } from '../base';
+import { ILogger } from '../logging-interface';
 import { KeyValueEntry } from './common';
 import { IReadOnlyResultMap, ResultMapForEachCb, ResultMapResultDetail } from './readonlyResultMap';
